after meeting coding
